[avalanche] Sign the avalanche proof
Summary:
In order to make it posible to update some parameters of a proof without having to sign the stakes again, we need to reduce the scope of what is signed by the stake. This means that we need to compensate this lack of stake signature score, and this diff creates a new proof signature to achieve that. The proof signs the limited proofid with the key corresponding to the master key, so that no one but the master key owner can update the proof without having the stake to sign again.

As a consequence to the master key change, there is one test case (Changing the master key affect ProofId) which is no longer relevant, as changing the key will raise a proof signature error and no longer a stake signature error, and ends up duplicated. This test case has been removed.

Ref T1676.

Depends on D10300.

Test Plan:
  ninja all check-all

Reviewers: #bitcoin_abc, PiRK, deadalnix

Reviewed By: #bitcoin_abc, PiRK, deadalnix

Subscribers: deadalnix

Maniphest Tasks: T1676

Differential Revision: https://reviews.bitcoinabc.org/D10140