[avalanche] Split block_register test in two cases
Summary:
The test is about the registering a vote on a block, but at the end it also checks that it's not possible to add a block for polling twice, which is a different feature. This diff extracts the last part to its own test case. This is a minor refactor that will make it easier to review the following test changes.

Ref T1854.

Test Plan:
  ninja check-avalanche-processor_tests

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Maniphest Tasks: T1854

Differential Revision: https://reviews.bitcoinabc.org/D10260