[avalanche] Introduce a test context in the processor tests
Summary:
This context object provides an interface which can be used for proofs, and make it possible to run the same test on both blocks and proofs where relevant thanks to the boost test case template feature. A single test is converted now to use the context in order to demonstrate the features, and others can be converted as well as needed.

Ref T1854.

Depends on D10260.

Test Plan:
  ninja check-avalanche-processor_tests

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Maniphest Tasks: T1854

Differential Revision: https://reviews.bitcoinabc.org/D10261