Fix the check-seeds.sh wrong evaluation of the live nodes
Summary:
The script count the number of live nodes vs the number of expected nodes, and returns the result as a percentage. I have been monitoring it recently after this percentage dropped below the warning threshold, and found this in the log:
  114% of testnet seeds appear to be online.

That's good news, but that's obviously wrong. The error comes from the bitcoin abc stop message sneaking to stdout and messing the count. Because we don't have so many testnet nodes to check, this can lead to this kind of error.

Test Plan:
  ./contrib/seeds/test-seeds.sh --testnet < ./contrib/seeds/nodes_test.txt > out.txt
The out.txt file should no longer contain the `Bitcoin ABC stopping` line

Reviewers: #bitcoin_abc, PiRK

Reviewed By: #bitcoin_abc, PiRK

Differential Revision: https://reviews.bitcoinabc.org/D10328