[avalanche] Convert the test utils to use proof shared pointers
Summary:
This is what is used almost everywhere in the API, so converting the test facilities removes the need for a lot of `std::make_shared<Proof>()`.

Note that `buildRandomProofRef` would be a better name after D10342, so I didn't rename the function in this diff.

Ref T1854.

Test Plan:
  ninja check

Reviewers: #bitcoin_abc, deadalnix

Reviewed By: #bitcoin_abc, deadalnix

Maniphest Tasks: T1854

Differential Revision: https://reviews.bitcoinabc.org/D10349